# Hoist all bounds checks out of loops