@jub0bs 2025-05-18
## Comparison to dgryski's implementation

### Before

```text
goos: darwin
goarch: amd64
pkg: github.com/agnivade/levenshtein
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                   │   dgryski    │               agniva                │
                   │    sec/op    │   sec/op     vs base                │
All/case=ASCII-8      335.6n ± 0%   163.1n ± 2%  -51.41% (p=0.000 n=20)
All/case=French-8     616.0n ± 0%   267.0n ± 2%  -56.66% (p=0.000 n=20)
All/case=Nordic-8    1184.0n ± 1%   429.3n ± 1%  -63.74% (p=0.000 n=20)
All/case=Tibetan-8    952.8n ± 2%   555.7n ± 1%  -41.67% (p=0.000 n=20)
geomean               694.9n        319.2n       -54.06%
```

### After

```text
goos: darwin
goarch: amd64
pkg: github.com/agnivade/levenshtein
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                   │   dgryski    │               agniva                │
                   │    sec/op    │   sec/op     vs base                │
All/case=ASCII-8      342.1n ± 1%   131.2n ± 3%  -61.63% (p=0.000 n=20)
All/case=French-8     628.1n ± 0%   221.9n ± 2%  -64.68% (p=0.000 n=20)
All/case=Nordic-8    1185.5n ± 1%   335.0n ± 2%  -71.74% (p=0.000 n=20)
All/case=Tibetan-8    974.5n ± 1%   493.4n ± 1%  -49.37% (p=0.000 n=20)
geomean               705.8n        263.4n       -62.68%
```

## Comparison to arbovm's implementation

### Before

```text
goos: darwin
goarch: amd64
pkg: github.com/agnivade/levenshtein
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                   │    arbovm    │               agniva                │
                   │    sec/op    │   sec/op     vs base                │
All/case=ASCII-8      356.8n ± 1%   163.1n ± 2%  -54.30% (p=0.000 n=20)
All/case=French-8     644.8n ± 0%   267.0n ± 2%  -58.60% (p=0.000 n=20)
All/case=Nordic-8    1208.5n ± 1%   429.3n ± 1%  -64.48% (p=0.000 n=20)
All/case=Tibetan-8    976.7n ± 3%   555.7n ± 1%  -43.10% (p=0.000 n=20)
geomean               721.9n        319.2n       -55.78%
```

### After

```text
goos: darwin
goarch: amd64
pkg: github.com/agnivade/levenshtein
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                   │    arbovm    │               agniva                │
                   │    sec/op    │   sec/op     vs base                │
All/case=ASCII-8      358.6n ± 1%   131.2n ± 3%  -63.40% (p=0.000 n=20)
All/case=French-8     653.3n ± 1%   221.9n ± 2%  -66.04% (p=0.000 n=20)
All/case=Nordic-8    1219.0n ± 1%   335.0n ± 2%  -72.52% (p=0.000 n=20)
All/case=Tibetan-8    980.0n ± 2%   493.4n ± 1%  -49.65% (p=0.000 n=20)
geomean               727.3n        263.4n       -63.79%
```