@jub0bs 2025-05-19
Can I change (as part of this PR) the names of the sub-benchmarks, in order for the results to be easier to analyse with benchstat?