@agnivade 2025-05-19
> I guess the performance gain largely depends on the two input strings: if they have a long common prefix and/or a long common suffix,

Yes. Could you add the additional benchmarks you added for the common prefix and suffixes? I don't see them in the PR.