@jub0bs 2025-05-19
> Yes. Could you add the additional benchmarks you added for the common prefix and suffixes? I don't see them in the PR.

I haven't added any new cases yet, but I can.