@agnivade 2025-05-19
My bad there. Somehow I forgot that the long lead and long trail cases already exist. Thanks!